+++ apps/client/src/pages/my/page/artist/artist-more.tsx
@@ -1,9 +1,10 @@
 import { useState } from 'react';
 import ArtistList from '@pages/my/components/artist/artist-list';
-import { useMyArtist } from '@pages/my/hooks/use-my-favorites';
+import { useSuspenseQuery } from '@tanstack/react-query';
 
 import { Header } from '@confeti/design-system';
 import { IcSwitch } from '@confeti/design-system/icons';
+import { USER_QUERY_OPTIONS } from '@shared/apis/user/user-queries';
 import {
   SORT_LABELS,
   SORT_OPTIONS,
@@ -14,7 +15,9 @@ import * as styles from './artist-more.css';
 
 const ArtistMore = () => {
   const [sortOption, setSortOption] = useState<SortOption>(SORT_OPTIONS.RECENT);
-  const { data } = useMyArtist(sortOption);
+  const { data } = useSuspenseQuery({
+    ...USER_QUERY_OPTIONS.MY_ARTISTS(sortOption),
+  });
 
   const toggleSort = () => {
     setSortOption((prev) =>
