+++ apps/client/src/pages/performance/components/performance-info/performance-info.tsx
@@ -1,5 +1,5 @@
 import { LikeButton } from '@confeti/design-system';
-import { useLikeMutation } from '@shared/hooks/use-like-mutation';
+import { useLikeMutation } from '@shared/hooks/queries/use-like-mutation';
 import { checkIsNotLoggedIn } from '@shared/utils/check-is-not-logged-in';
 import { formatDate } from '@shared/utils/format-date';
